requirements.txt
ansible==5.5.0
ansible-core==2.12.4
apache-libcloud==3.4.1
argcomplete==2.0.0
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.4
cachetools==5.0.0
certifi==2022.12.7
chardet==4.0.0
click==8.0.3
colorama==0.4.5
comm==0.1.3
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu1
dnspython==2.2.1
duplicity==0.8.22
exceptiongroup==1.0.4
executing==1.2.0
fasteners==0.17.3
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
h11==0.14.0
httplib2==0.20.4
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.0.3
jmespath==1.0.1
jupyter_client==8.1.0
jupyter_core==5.3.0
keyring==23.8.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.22.0
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==21.3
paramiko==2.10.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.4.0
pykerberos==1.1.14
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-apt==2.4.0+22.10
python-dateutil==2.8.2
python-debian==0.1.46
pytz==2022.2.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
reportlab==3.6.11
requests==2.27.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
resolvelib==0.8.1
rsa==4.8
SecretStorage==3.3.3
selenium==4.7.2
selinux==3.4
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
stack-data==0.6.2
systemd-python==235
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
